+++ DfE.GIAP.All/src/DfE.GIAP.Core/Common/CrossCutting/UniquePupilNumberValidator.cs
@@ -1,6 +1,6 @@
 ﻿using System.Text.RegularExpressions;
 
-namespace DfE.GIAP.Core.MyPupils.Domain.ValueObjects;
+namespace DfE.GIAP.Core.Common.CrossCutting;
 public static class UniquePupilNumberValidator
 {
     // see GOV.UK spec
