+++ DfE.GIAP.All/tests/DfE.GIAP.SharedTests/TestDoubles/UniquePupilNumberTestDoubles.cs
@@ -1,4 +1,5 @@
-﻿using DfE.GIAP.Core.MyPupils.Domain.ValueObjects;
+﻿using DfE.GIAP.Core.Common.CrossCutting;
+using DfE.GIAP.Core.MyPupils.Domain.ValueObjects;
 using Fare;
 
 namespace DfE.GIAP.SharedTests.TestDoubles;
